Bridge/requirements.txt
paho-mqtt
pyserial
requests
